#4 Implemented possibility to have letters in the version number in plugin.yaml

+++ src/main/java/de/bywambo/chattodiscord/ChatToDiscord.java
@@ -26,7 +26,7 @@ public class ChatToDiscord extends JavaPlugin {
             return;
 
         try {
-            UpdateChecker.checkForUpdate(Float.valueOf(this.getDescription().getVersion()));
+            UpdateChecker.checkForUpdate(this.getDescription().getVersion());
         } catch (IOException e) {
             Bukkit.getLogger().info("Couldn't check version on GitHub. Please check your internet connection");
         }
